added different project for each layer
