+++ Library/Api/Program.cs
@@ -1,0 +1,43 @@
+using Microsoft.EntityFrameworkCore;
+
+using Application.Services;
+using Domain.Interfaces;
+using Infrastructure.Persistence;
+using Infrastructure.Repositories;
+
+var builder = WebApplication.CreateBuilder(args);
+
+builder.AddServiceDefaults();
+
+builder.AddNpgsqlDbContext<AppDbContext>(connectionName: "DefaultConnection");
+
+builder.Services.AddScoped<ICustomerRepository, CustomerRepository>();
+builder.Services.AddScoped<IBookRepository, BookRepository>();
+builder.Services.AddScoped<IBorrowRecordRepository, BorrowRecordRepository>();
+
+builder.Services.AddScoped<AnalyticsService>();
+
+builder.Services.AddControllers();
+builder.Services.AddEndpointsApiExplorer();
+builder.Services.AddSwaggerGen();
+
+var app = builder.Build();
+
+using (var scope = app.Services.CreateScope())
+{
+    var db = scope.ServiceProvider.GetRequiredService<AppDbContext>();
+    db.Database.Migrate();
+}
+
+if (app.Environment.IsDevelopment())
+{
+    app.UseSwagger();
+    app.UseSwaggerUI();
+}
+
+app.UseHttpsRedirection();
+app.UseAuthorization();
+app.MapDefaultEndpoints();
+app.MapControllers();
+
+app.Run();
