+++ Library/Domain/Enums/Publisher.cs
@@ -1,0 +1,18 @@
+﻿namespace Domain.Enums;
+
+/// <summary>
+/// Represents the publisher of a book.
+/// </summary>
+public enum Publisher 
+{
+    NorthernWord,
+    BookMosaic,
+    LiteraryWind,
+    GoldenPage,
+    EchoOfThought,
+    TheBinding,
+    IntellectPublishing,
+    WhiteLine,
+    BeaconPress,
+    NewEraPublishing
+}
