+++ Library/Domain/Enums/PublishingType.cs
@@ -1,0 +1,13 @@
+﻿namespace Domain.Enums;
+
+/// <summary>
+/// Represents the type or format of a book's publishing.
+/// </summary>
+public enum PublishingType 
+{
+    Hardcover,
+    Paperback,
+    Ebook,
+    Audiobook,
+    LimitedEdition
+}
