+++ library/library.AppHost/Properties/launchSettings.json
@@ -5,24 +5,24 @@
       "commandName": "Project",
       "dotnetRunMessages": true,
       "launchBrowser": true,
-      "applicationUrl": "https://localhost:17243;http://localhost:15033",
+      "applicationUrl": "https://localhost:17148;http://localhost:15188",
       "environmentVariables": {
         "ASPNETCORE_ENVIRONMENT": "Development",
         "DOTNET_ENVIRONMENT": "Development",
-        "ASPIRE_DASHBOARD_OTLP_ENDPOINT_URL": "https://localhost:21236",
-        "ASPIRE_RESOURCE_SERVICE_ENDPOINT_URL": "https://localhost:22008"
+        "DOTNET_DASHBOARD_OTLP_ENDPOINT_URL": "https://localhost:21054",
+        "DOTNET_RESOURCE_SERVICE_ENDPOINT_URL": "https://localhost:22066"
       }
     },
     "http": {
       "commandName": "Project",
       "dotnetRunMessages": true,
       "launchBrowser": true,
-      "applicationUrl": "http://localhost:15033",
+      "applicationUrl": "http://localhost:15188",
       "environmentVariables": {
         "ASPNETCORE_ENVIRONMENT": "Development",
         "DOTNET_ENVIRONMENT": "Development",
-        "ASPIRE_DASHBOARD_OTLP_ENDPOINT_URL": "http://localhost:19154",
-        "ASPIRE_RESOURCE_SERVICE_ENDPOINT_URL": "http://localhost:20258"
+        "DOTNET_DASHBOARD_OTLP_ENDPOINT_URL": "http://localhost:19079",
+        "DOTNET_RESOURCE_SERVICE_ENDPOINT_URL": "http://localhost:20078"
       }
     }
   }
