--- library/library/Domain/Enums/Publisher.cs
@@ -1,18 +1,0 @@
-﻿namespace library.Domain.Enums;
-
-/// <summary>
-/// Represents the publisher of a book.
-/// </summary>
-public enum Publisher 
-{
-    NorthernWord,
-    BookMosaic,
-    LiteraryWind,
-    GoldenPage,
-    EchoOfThought,
-    TheBinding,
-    IntellectPublishing,
-    WhiteLine,
-    BeaconPress,
-    NewEraPublishing
-}
