--- library/library/Domain/Enums/PublishingType.cs
@@ -1,13 +1,0 @@
-﻿namespace library.Domain.Enums;
-
-/// <summary>
-/// Represents the type or format of a book's publishing.
-/// </summary>
-public enum PublishingType 
-{
-    Hardcover,
-    Paperback,
-    Ebook,
-    Audiobook,
-    LimitedEdition
-}
