+++ library/library/appsettings.Development.json
@@ -1,4 +1,5 @@
 {
+  "DetailedErrors": true,
   "Logging": {
     "LogLevel": {
       "Default": "Information",
